working now..

+++ S1/p1_lr.c
@@ -277,7 +277,7 @@ void
 lr_invert(pmatrix a){
 
     int i,j,k;
-    int sum;
+    double sum;
     int ld = a->ld;
     int n = a->rows;
     double *aa = a->a;
@@ -285,13 +285,14 @@ lr_invert(pmatrix a){
     // inversion of L
     for (i = 0; i < n; i++) {
         for (j = 0; j < i; j++) {
-            sum = (int) aa[i + j*ld];
+            sum = aa[i + j*ld];
             for (k = j+1; k < i; k++) {
                 sum += aa[i + k*ld] * aa[k + j*ld];
             }
             aa[i + j*ld] = -1 * sum;
         }
     }
+
 
     // inversion of R
     for (i = n-1; i >= 0; i--) {
